backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
requests==2.26.0
python-dotenv==0.19.2
SQLAlchemy==1.4.27
PyMySQL==1.0.2
Werkzeug==2.0.3